requirements.txt
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
easyocr==1.7.2
fastapi==0.95.2
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
h11==0.14.0
idna==3.10
imageio==2.35.1
Jinja2==3.1.5
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.3
numpy==1.24.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pipdeptree==2.24.0
psutil==6.1.1
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==1.10.9
pydantic_core==2.27.2
pyparsing==3.1.4
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.13.1
tifffile==2023.7.10
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.65
ultralytics-thop==2.0.14
urllib3==2.2.3
uvicorn==0.33.0